+++ src/main/java/me/mrCookieSlime/Slimefun/holograms/HologramProjectorHologram.java
@@ -26,7 +26,9 @@ public final class HologramProjectorHologram {
 		Location l = new Location(projector.getWorld(), projector.getX() + 0.5, projector.getY() + offset, projector.getZ() + 0.5);
 		
 		for (Entity n : l.getChunk().getEntities()) {
-			if (n instanceof ArmorStand && n.getCustomName() != null && n.getCustomName().equals(nametag) && l.distanceSquared(n.getLocation()) < 0.4D) return (ArmorStand) n;
+			if (n instanceof ArmorStand && n.getCustomName() != null && n.getCustomName().equals(nametag) && l.distanceSquared(n.getLocation()) < 0.4D) {
+				return (ArmorStand) n;
+			}
 		}
 		
 		if (!createIfNoneExists) {