+++ src/main/java/me/mrCookieSlime/Slimefun/holograms/ReactorHologram.java
@@ -1,12 +1,11 @@
 package me.mrCookieSlime.Slimefun.holograms;
 
-import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Entity;
 
-import me.mrCookieSlime.Slimefun.SlimefunPlugin;
+import me.mrCookieSlime.Slimefun.api.Slimefun;
 
 public final class ReactorHologram {
 	
@@ -16,7 +15,9 @@ public final class ReactorHologram {
         Location l = new Location(reactor.getWorld(), reactor.getX() + 0.5, reactor.getY() + 0.7, reactor.getZ() + 0.5);
 
         for (Entity n : l.getChunk().getEntities()) {
-            if (n instanceof ArmorStand && l.distanceSquared(n.getLocation()) < 0.4D) return (ArmorStand) n;
+            if (n instanceof ArmorStand && l.distanceSquared(n.getLocation()) < 0.4D) {
+            	return (ArmorStand) n;
+            }
         }
         
         if (!createIfNoneExists) return null;
@@ -27,18 +28,11 @@ public final class ReactorHologram {
         return hologram;
     }
     
-    public static void update(final Location l, final String name) {
-		Bukkit.getScheduler().scheduleSyncDelayedTask(SlimefunPlugin.instance, () -> {
+    public static void update(Location l, String name) {
+		Slimefun.runSync(() -> {
 			ArmorStand hologram = getArmorStand(l, true);
 			if (!hologram.isCustomNameVisible()) hologram.setCustomNameVisible(true);
 			hologram.setCustomName(ChatColor.translateAlternateColorCodes('&', name));
 		});
 	}
-
-    public static void remove(Location l) {
-    	Bukkit.getScheduler().scheduleSyncDelayedTask(SlimefunPlugin.instance, () -> {
-            ArmorStand hologram = getArmorStand(l, false);
-            if (hologram != null) hologram.remove();
-		});
-    }
 }