+++ src/main/java/me/mrCookieSlime/Slimefun/holograms/SimpleHologram.java
@@ -1,6 +1,5 @@
 package me.mrCookieSlime.Slimefun.holograms;
 
-import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.block.Block;
@@ -8,21 +7,21 @@ import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
 
-import me.mrCookieSlime.Slimefun.SlimefunPlugin;
+import me.mrCookieSlime.Slimefun.api.Slimefun;
 
 public final class SimpleHologram {
 	
 	private SimpleHologram() {}
 	
-	public static void update(final Block b, final String name) {
-		Bukkit.getScheduler().scheduleSyncDelayedTask(SlimefunPlugin.instance, () -> {
+	public static void update(Block b, String name) {
+		Slimefun.runSync(() -> {
 			ArmorStand hologram = getArmorStand(b, true);
 			hologram.setCustomName(ChatColor.translateAlternateColorCodes('&', name));
 		});
 	}	
 	
-	public static void remove(final Block b) {
-		Bukkit.getScheduler().scheduleSyncDelayedTask(SlimefunPlugin.instance, () -> {
+	public static void remove(Block b) {
+		Slimefun.runSync(() -> {
 			ArmorStand hologram = getArmorStand(b, false);
 			if (hologram != null) hologram.remove();
 		});
@@ -32,7 +31,9 @@ public final class SimpleHologram {
 		Location l = new Location(b.getWorld(), b.getX() + 0.5, b.getY() + 0.7F, b.getZ() + 0.5);
 		
 		for (Entity n : l.getChunk().getEntities()) {
-			if (n instanceof ArmorStand && n.getCustomName() != null && l.distanceSquared(n.getLocation()) < 0.4D) return (ArmorStand) n;
+			if (n instanceof ArmorStand && n.getCustomName() != null && l.distanceSquared(n.getLocation()) < 0.4D) {
+				return (ArmorStand) n;
+			}
 		}
 		
 		if (!createIfNoneExists) return null;